Revert "feat: Add directional resize buttons to code editor's resize control …"
This reverts commit b4f0b2f95fbe0508ee06020e8a930caf1a1ede21.
