+++ pages/code-editor/base-props.ts
@@ -35,7 +35,4 @@ export const i18nStrings: CodeEditorProps.I18nStrings = {
   preferencesModalThemeFilteringAriaLabel: 'Filter themes',
   preferencesModalThemeFilteringClearAriaLabel: 'Clear',
   preferencesModalThemeFilteringPlaceholder: 'Filter themes',
-
-  resizeHandleAriaLabel: 'Resize handle',
-  resizeHandleTooltipText: 'Drag or select to resize',
 };
