+++ src/code-editor/interfaces.ts
@@ -167,9 +167,6 @@ export namespace CodeEditorProps {
     preferencesModalThemeFilteringPlaceholder?: string;
     preferencesModalThemeFilteringAriaLabel?: string;
     preferencesModalThemeFilteringClearAriaLabel?: string;
-
-    resizeHandleAriaLabel?: string;
-    resizeHandleTooltipText?: string;
   }
   export interface ResizeDetail {
     height: number;
