+++ src/code-editor/pane.tsx
@@ -26,8 +26,6 @@ interface PaneProps {
 
   cursorPositionLabel?: (row: number, column: number) => string;
   closeButtonAriaLabel?: string;
-  handleAriaLabel?: string;
-  handleTooltipText?: string;
 
   onClose: () => void;
   onAnnotationClick: (annotation: Ace.Annotation) => void;
@@ -45,8 +43,6 @@ export const Pane = ({
   onAnnotationClear,
   cursorPositionLabel,
   closeButtonAriaLabel,
-  handleAriaLabel,
-  handleTooltipText,
 }: PaneProps) => {
   const [paneHeight, setPaneHeight] = useState(MIN_HEIGHT);
   const listRef = useRef<HTMLTableSectionElement>(null);
@@ -92,13 +88,7 @@ export const Pane = ({
 
   return (
     <div id={id} className={styles.pane} onKeyDown={onEscKeyDown} role="tabpanel" aria-labelledby={ariaLabelledBy}>
-      <ResizableBox
-        height={paneHeight}
-        minHeight={MIN_HEIGHT}
-        onResize={newHeight => setPaneHeight(newHeight)}
-        handleAriaLabel={handleAriaLabel}
-        handleTooltipText={handleTooltipText}
-      >
+      <ResizableBox height={paneHeight} minHeight={MIN_HEIGHT} onResize={newHeight => setPaneHeight(newHeight)}>
         <FocusLock className={styles['focus-lock']} autoFocus={true} restoreFocus={true}>
           <div className={styles.pane__list} tabIndex={-1}>
             <table className={styles.pane__table} role="presentation">
