+++ src/code-editor/resizable-box/__tests__/resizable-box.test.tsx
@@ -4,10 +4,7 @@ import React, { useState } from 'react';
 import { fireEvent, render } from '@testing-library/react';
 
 import { ResizableBox, ResizeBoxProps } from '../../../../lib/components/code-editor/resizable-box';
-import { PointerEventMock } from '../../../../lib/components/internal/utils/pointer-events-mock';
 
-import dragHandleStyles from '../../../../lib/components/internal/components/drag-handle/styles.css.js';
-import dragHandleWrapperStyles from '../../../../lib/components/internal/components/drag-handle-wrapper/styles.css.js';
 import styles from '../../../../lib/components/code-editor/resizable-box/styles.selectors.js';
 
 const defaultProps: ResizeBoxProps = {
@@ -22,23 +19,8 @@ function findBox() {
 }
 
 function findHandle() {
-  return document.querySelector(`.${dragHandleStyles.handle}`)!;
+  return document.querySelector(`.${styles['resizable-box-handle']}`)!;
 }
-
-function findDirectionButton(direction: 'block-start' | 'block-end') {
-  return document.querySelector(
-    `.${dragHandleWrapperStyles[`direction-button-wrapper-${direction}`]} .${dragHandleWrapperStyles['direction-button']}`
-  )!;
-}
-
-beforeAll(() => {
-  (window as any).PointerEvent ??= PointerEventMock;
-});
-
-test('Provides ariaLabel to the handle', () => {
-  render(<ResizableBox {...defaultProps} handleAriaLabel="Resize handle" />);
-  expect(findHandle()).toHaveAccessibleName('Resize handle');
-});
 
 test('Height is controlled', () => {
   const { rerender } = render(<ResizableBox {...defaultProps} />);
@@ -47,58 +29,23 @@ test('Height is controlled', () => {
   expect(findBox()).toHaveStyle({ height: '200px' });
 });
 
-test('Emits resize events on mouse resizing', () => {
+test('Emits resize events', () => {
   const onResize = jest.fn();
   render(<ResizableBox {...defaultProps} onResize={onResize} />);
-  fireEvent.pointerDown(findHandle());
-  fireEvent.pointerMove(document.body, { clientY: 80 });
+  fireEvent.mouseDown(findHandle());
+  fireEvent.mouseMove(document.body, { clientY: 80 });
   expect(onResize).toHaveBeenCalledWith(80);
   onResize.mockClear();
-  fireEvent.pointerMove(document.body, { clientY: 150 });
+  fireEvent.mouseMove(document.body, { clientY: 150 });
   expect(onResize).toHaveBeenCalledWith(150);
-});
-
-test('Emits resize events on keyboard resizing', () => {
-  const onResize = jest.fn();
-  render(<ResizableBox {...defaultProps} height={100} onResize={onResize} />);
-  // Arrow keys resize by 20px
-  fireEvent.keyDown(findHandle(), { key: 'ArrowDown' });
-  expect(onResize).toHaveBeenCalledWith(120);
-
-  onResize.mockClear();
-  fireEvent.keyDown(findHandle(), { key: 'ArrowRight' });
-  expect(onResize).toHaveBeenCalledWith(120);
-
-  onResize.mockClear();
-  fireEvent.keyDown(findHandle(), { key: 'ArrowUp' });
-  expect(onResize).toHaveBeenCalledWith(80);
-
-  onResize.mockClear();
-  fireEvent.keyDown(findHandle(), { key: 'ArrowLeft' });
-  expect(onResize).toHaveBeenCalledWith(80);
-});
-
-test('Emits resize events when direction buttons are pressed', () => {
-  const onResize = jest.fn();
-  render(<ResizableBox {...defaultProps} height={100} onResize={onResize} />);
-  // Open direction buttons
-  fireEvent.pointerDown(findHandle());
-  fireEvent.pointerUp(findHandle());
-
-  fireEvent.click(findDirectionButton('block-start'));
-  expect(onResize).toHaveBeenCalledWith(80);
-
-  onResize.mockClear();
-  fireEvent.click(findDirectionButton('block-end'));
-  expect(onResize).toHaveBeenCalledWith(120);
 });
 
 test('Includes offsets into size calculation', () => {
   const onResize = jest.fn();
   render(<ResizableBox {...defaultProps} onResize={onResize} />);
   findBox().getBoundingClientRect = () => ({ top: 50, bottom: 100 }) as DOMRect;
-  fireEvent.pointerDown(findHandle(), { clientY: 95 });
-  fireEvent.pointerMove(document.body, { clientY: 130 });
+  fireEvent.mouseDown(findHandle(), { clientY: 95 });
+  fireEvent.mouseMove(document.body, { clientY: 130 });
   // +5 offset for mouseDown position (100 - 95)
   // -50 for container top offset
   expect(onResize).toHaveBeenCalledWith(85); // 85 = 130 + 5 - 50
@@ -106,34 +53,34 @@ test('Includes offsets into size calculation', () => {
 
 test('Disables selection while drag is active', () => {
   render(<ResizableBox {...defaultProps} />);
-  fireEvent.pointerDown(findHandle());
+  fireEvent.mouseDown(findHandle());
   expect(document.body).toHaveClass(styles['resize-active']);
-  fireEvent.pointerUp(document.body);
+  fireEvent.mouseUp(document.body);
   expect(document.body).not.toHaveClass(styles['resize-active']);
 });
 
 test('Does not respond to events after cursor release', () => {
   const onResize = jest.fn();
   render(<ResizableBox {...defaultProps} onResize={onResize} />);
-  fireEvent.pointerDown(findHandle());
-  fireEvent.pointerUp(document.body);
-  fireEvent.pointerMove(document.body, { clientY: 80 });
+  fireEvent.mouseDown(findHandle());
+  fireEvent.mouseUp(document.body);
+  fireEvent.mouseMove(document.body, { clientY: 80 });
   expect(onResize).not.toHaveBeenCalled();
 });
 
 test('Has no effect on other non-left mouse button clicks', () => {
   const onResize = jest.fn();
   render(<ResizableBox {...defaultProps} minHeight={50} onResize={onResize} />);
-  fireEvent.pointerDown(findHandle(), { pointerType: 'mouse', button: 1 });
-  fireEvent.pointerMove(document.body, { clientY: 30 });
+  fireEvent.mouseDown(findHandle(), { button: 1 });
+  fireEvent.mouseMove(document.body, { clientY: 30 });
   expect(onResize).not.toHaveBeenCalled();
 });
 
 test('Does not allow to go beyond min size', () => {
   const onResize = jest.fn();
   render(<ResizableBox {...defaultProps} minHeight={50} onResize={onResize} />);
-  fireEvent.pointerDown(findHandle());
-  fireEvent.pointerMove(document.body, { clientY: 30 });
+  fireEvent.mouseDown(findHandle());
+  fireEvent.mouseMove(document.body, { clientY: 30 });
   expect(onResize).toHaveBeenCalledWith(50);
 });
 
@@ -144,8 +91,8 @@ test('Does not cause state warnings after unmounting', () => {
     return <ResizableBox {...defaultProps} height={height} onResize={height => setHeight(height)} />;
   }
   const { rerender } = render(<Stateful />);
-  fireEvent.pointerDown(findHandle());
+  fireEvent.mouseDown(findHandle());
   rerender(<></>);
-  fireEvent.pointerMove(document.body, { clientY: 30 });
+  fireEvent.mouseMove(document.body, { clientY: 30 });
   expect(consoleMock).not.toHaveBeenCalled();
 });
