+++ src/code-editor/resizable-box/index.tsx
@@ -4,53 +4,26 @@ import React, { useEffect, useRef, useState } from 'react';
 
 import { useStableCallback } from '@cloudscape-design/component-toolkit/internal';
 
-import InternalDragHandle from '../../internal/components/drag-handle';
-
 import styles from './styles.css.js';
-
-const KEYBOARD_STEP_SIZE = 20;
 
 export interface ResizeBoxProps {
   children: React.ReactNode;
   height: number;
   minHeight: number;
   onResize: (newHeight: number) => void;
-
-  handleAriaLabel?: string;
-  handleTooltipText?: string;
 }
 
-export function ResizableBox({
-  children,
-  height,
-  minHeight,
-  onResize,
-  handleAriaLabel,
-  handleTooltipText,
-}: ResizeBoxProps) {
+export function ResizableBox({ children, height, minHeight, onResize }: ResizeBoxProps) {
   const [dragOffset, setDragOffset] = useState<null | number>(null);
   const onResizeStable = useStableCallback(onResize);
   const containerRef = useRef<HTMLDivElement>(null);
 
-  const onPointerDown: React.PointerEventHandler = event => {
-    if ((event.pointerType === 'mouse' && event.button !== 0) || !containerRef.current) {
+  const onMouseDown: React.MouseEventHandler = event => {
+    if (event.button !== 0 || !containerRef.current) {
       return;
     }
     const containerBottom = containerRef.current.getBoundingClientRect().bottom;
     setDragOffset(containerBottom - event.clientY);
-  };
-
-  const onKeyDown: React.KeyboardEventHandler = event => {
-    switch (event.key) {
-      case 'ArrowDown':
-      case 'ArrowRight':
-        onResizeStable(height + KEYBOARD_STEP_SIZE);
-        break;
-      case 'ArrowUp':
-      case 'ArrowLeft':
-        onResizeStable(Math.max(height - KEYBOARD_STEP_SIZE, minHeight));
-        break;
-    }
   };
 
   useEffect(() => {
@@ -59,56 +32,28 @@ export function ResizableBox({
     }
     const container = containerRef.current;
 
-    const onPointerMove = (event: PointerEvent) => {
+    const onMouseMove = (event: MouseEvent) => {
       const { top } = container.getBoundingClientRect();
       const cursor = event.clientY;
       onResizeStable(Math.max(cursor + dragOffset - top, minHeight));
     };
-    const onPointerUp = () => {
+    const onMouseUp = () => {
       setDragOffset(null);
     };
     const controller = new AbortController();
     document.body.classList.add(styles['resize-active']);
-    container.classList.add(styles['cursor-active']);
-    document.addEventListener('pointermove', onPointerMove, { signal: controller.signal });
-    document.addEventListener('pointerup', onPointerUp, { signal: controller.signal });
+    document.addEventListener('mousemove', onMouseMove, { signal: controller.signal });
+    document.addEventListener('mouseup', onMouseUp, { signal: controller.signal });
     return () => {
       controller.abort();
       document.body.classList.remove(styles['resize-active']);
-      container.classList.remove(styles['cursor-active']);
     };
   }, [dragOffset, minHeight, onResizeStable]);
 
   return (
     <div ref={containerRef} className={styles['resizable-box']} style={{ height }}>
       {children}
-
-      <div className={styles['resizable-box-handle']}>
-        <InternalDragHandle
-          ariaLabel={handleAriaLabel}
-          variant="resize-area"
-          // Provide an arbitrary large value to valueMax since the editor can be
-          // resized to be infinitely large.
-          ariaValue={{ valueMin: minHeight, valueMax: 1000000, valueNow: height }}
-          tooltipText={handleTooltipText}
-          onPointerDown={onPointerDown}
-          onKeyDown={onKeyDown}
-          directions={{
-            'block-start': height > minHeight ? 'active' : 'disabled',
-            'block-end': 'active',
-          }}
-          onDirectionClick={direction => {
-            switch (direction) {
-              case 'block-end':
-                onResizeStable(height + KEYBOARD_STEP_SIZE);
-                break;
-              case 'block-start':
-                onResizeStable(Math.max(height - KEYBOARD_STEP_SIZE, minHeight));
-                break;
-            }
-          }}
-        />
-      </div>
+      <span className={styles['resizable-box-handle']} onMouseDown={onMouseDown} />
     </div>
   );
 }
