+++ src/i18n/messages/all.es.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "Filtrar temas",
     "i18nStrings.preferencesModalLightThemes": "Temas claros",
     "i18nStrings.preferencesModalDarkThemes": "Temas oscuros",
-    "i18nStrings.cursorPositionAriaLabel": "Cursor en fila {row}",
-    "i18nStrings.resizeHandleAriaLabel": "Cambiar el tamaño del controlador",
-    "i18nStrings.resizeHandleTooltipText": "Arrastre o seleccione para cambiar el tamaño"
+    "i18nStrings.cursorPositionAriaLabel": "Cursor en fila {row}"
   },
   "collection-preferences": {
     "title": "Preferencias",
