+++ src/i18n/messages/all.fr.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "Filtrer les thèmes",
     "i18nStrings.preferencesModalLightThemes": "Thèmes clairs",
     "i18nStrings.preferencesModalDarkThemes": "Thèmes sombres",
-    "i18nStrings.cursorPositionAriaLabel": "Curseur sur la ligne {row}",
-    "i18nStrings.resizeHandleAriaLabel": "Poignée de redimensionnement",
-    "i18nStrings.resizeHandleTooltipText": "Faites glisser ou sélectionnez pour redimensionner"
+    "i18nStrings.cursorPositionAriaLabel": "Curseur sur la ligne {row}"
   },
   "collection-preferences": {
     "title": "Préférences",
