+++ src/i18n/messages/all.id.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "Filter tema",
     "i18nStrings.preferencesModalLightThemes": "Tema terang",
     "i18nStrings.preferencesModalDarkThemes": "Tema gelap",
-    "i18nStrings.cursorPositionAriaLabel": "Kursor pada baris {row}",
-    "i18nStrings.resizeHandleAriaLabel": "Tuas pengubah ukuran",
-    "i18nStrings.resizeHandleTooltipText": "Seret atau pilih untuk mengubah ukuran"
+    "i18nStrings.cursorPositionAriaLabel": "Kursor pada baris {row}"
   },
   "collection-preferences": {
     "title": "Preferensi",
