+++ src/i18n/messages/all.pt-BR.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "Filtrar temas",
     "i18nStrings.preferencesModalLightThemes": "Temas claros",
     "i18nStrings.preferencesModalDarkThemes": "Temas escuros",
-    "i18nStrings.cursorPositionAriaLabel": "Cursor na linha {row}",
-    "i18nStrings.resizeHandleAriaLabel": "Redimensionar alça",
-    "i18nStrings.resizeHandleTooltipText": "Arraste ou selecione para redimensionar"
+    "i18nStrings.cursorPositionAriaLabel": "Cursor na linha {row}"
   },
   "collection-preferences": {
     "title": "Preferências",
