+++ src/i18n/messages/all.tr.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "Temaları filtrele",
     "i18nStrings.preferencesModalLightThemes": "Açık temalar",
     "i18nStrings.preferencesModalDarkThemes": "Koyu temalar",
-    "i18nStrings.cursorPositionAriaLabel": "İmleç {row}. satırda",
-    "i18nStrings.resizeHandleAriaLabel": "Yeniden boyutlandırma tutamacı",
-    "i18nStrings.resizeHandleTooltipText": "Yeniden boyutlandırmak için sürükleyin veya seçin"
+    "i18nStrings.cursorPositionAriaLabel": "İmleç {row}. satırda"
   },
   "collection-preferences": {
     "title": "Tercihler",
