+++ src/i18n/messages/all.zh-TW.json
@@ -88,9 +88,7 @@
     "i18nStrings.preferencesModalThemeFilteringPlaceholder": "篩選主題",
     "i18nStrings.preferencesModalLightThemes": "淺色佈景主題",
     "i18nStrings.preferencesModalDarkThemes": "深色佈景主題",
-    "i18nStrings.cursorPositionAriaLabel": "列 {row} 上的游標",
-    "i18nStrings.resizeHandleAriaLabel": "調整大小控點",
-    "i18nStrings.resizeHandleTooltipText": "拖曳或選取以調整大小"
+    "i18nStrings.cursorPositionAriaLabel": "列 {row} 上的游標"
   },
   "collection-preferences": {
     "title": "偏好設定",
