+++ style-dictionary/utils/token-names.ts
@@ -249,10 +249,6 @@ export type ColorsTokenName =
   | 'colorBackgroundButtonPrimaryDefault'
   | 'colorBackgroundButtonPrimaryDisabled'
   | 'colorBackgroundButtonPrimaryHover'
-  | 'colorBackgroundDirectionButtonActive'
-  | 'colorBackgroundDirectionButtonDefault'
-  | 'colorBackgroundDirectionButtonDisabled'
-  | 'colorBackgroundDirectionButtonHover'
   | 'colorBackgroundCalendarCurrentDate'
   | 'colorBackgroundCellShaded'
   | 'colorBackgroundCodeEditorGutterActiveLineDefault'
@@ -410,8 +406,6 @@ export type ColorsTokenName =
   | 'colorTextButtonPrimaryActive'
   | 'colorTextButtonPrimaryDefault'
   | 'colorTextButtonPrimaryHover'
-  | 'colorTextDirectionButtonDefault'
-  | 'colorTextDirectionButtonDisabled'
   | 'colorTextCalendarDateHover'
   | 'colorTextCalendarDateSelected'
   | 'colorTextCalendarMonth'
