+++ style-dictionary/visual-refresh/colors.ts
@@ -26,12 +26,6 @@ const tokens: StyleDictionary.ColorsDictionary = {
   colorBackgroundButtonPrimaryDefault: '{colorBorderButtonNormalDefault}',
   colorBackgroundButtonPrimaryDisabled: { light: '{colorGrey200}', dark: '{colorGrey700}' },
   colorBackgroundButtonPrimaryHover: '{colorBorderButtonNormalHover}',
-  colorBackgroundDirectionButtonActive: '{colorGrey700}',
-  colorBackgroundDirectionButtonDefault: '{colorGrey600}',
-  colorBackgroundDirectionButtonDisabled: { light: '{colorGrey200}', dark: '{colorGrey700}' },
-  colorBackgroundDirectionButtonHover: '{colorGrey650}',
-  colorTextDirectionButtonDefault: '{colorWhite}',
-  colorTextDirectionButtonDisabled: '{colorTextInteractiveDisabled}',
   colorBackgroundCalendarCurrentDate: { light: '{colorGrey150}', dark: '{colorGrey650}' },
   colorBackgroundCellShaded: { light: '{colorGrey125}', dark: '{colorGrey750}' },
   colorBackgroundCodeEditorGutterActiveLineDefault: { light: '{colorGrey550}', dark: '{colorGrey500}' },
